Add PKCS12 HTTPS cert option

+++ start-winslow.sh
@@ -1,16 +1,22 @@
 #!/bin/bash
 
-#IMAGE="winslow"
-IMAGE="repo.itd-intern.de/winslow/node"
 HTTP="80"
-PARAMS="$@"
+HTTPS="443"
+#NODE_TYPE="executor"
+NODE_TYPE="observer"
 STORAGE_TYPE="nfs"
 STORAGE_PATH="pc973l:/home/mi7wa6/mec-view/winslow/nfs-export"
+KEYSTORE_PATH_PKCS12="/root/winslow.itd-intern.de.p12"
+ADDITIONAL="-p 446:4646 -e WINSLOW_DEV_ENV=true -e WINSLOW_DEV_REMOTE_USER=michael"
+
+PARAMS="$@"
+#IMAGE="winslow"
+IMAGE="repo.itd-intern.de/winslow/node"
+NODE_NAME="$(hostname)"
+CONTAINER_NAME="winslow"
+GPUS="$(ls /dev/ | grep -i nvidia | wc -l)"
 WORKDIR="/winslow/"
 
-NODE_NAME="$(hostname)"
-ADDITIONAL="-p 446:4646 -e WINSLOW_DEV_ENV=true -e WINSLOW_DEV_REMOTE_USER=michael"
-GPUS="$(ls /dev/ | grep -i nvidia | wc -l)"
 SUDO=""
 
 if [ "$(id -u)" -ne 0 ] && [ "$(id --name -G | grep -i docker | wc -l)" -eq 0 ]; then
@@ -18,6 +24,10 @@ if [ "$(id -u)" -ne 0 ] && [ "$(id --name -G | grep -i docker | wc -l)" -eq 0 ];
 fi
 
 $SUDO docker pull $IMAGE
+
+if [ "$KEYSTORE_PATH_PKCS12" != "" ]; then
+    ADDITIONAL="$ADDITIONAL -p $HTTPS:8080 -v $KEYSTORE_PATH_PKCS12:/keystore.p12:ro -e SERVER_KEY_ALIAS=winslow.itd-intern.de -e SERVER_SSL_KEY_STORE_TYPE=PKCS12 -e SERVER_SSL_KEY_STORE=file:/keystore.p12 -e SECURITY_REQUIRE_SSL=true -e SERVER_SSL_KEY_STORE_PASSWORD="
+fi
 
 
 echo ""
@@ -40,21 +50,23 @@ echo ""
 
 sleep 1
 
-GPU_FLAG=""
-
-if [ "$GPUS" -gt 0 ]; then
-    GPU_FLAG="--gpus all"
+if [ $($SUDO docker ps --filter "name=$CONTAINER_NAME" | wc -l) -gt 1 ]; then
+    echo " ::::: Stopping already running Winslow instance"
+    $SUDO docker stop "$CONTAINER_NAME" > /dev/null && echo "  :::: Done" || (echo " :::: Failed"; exit 1)
 fi
+
 
 echo " ::::: Starting Winslow Container now"
 $SUDO docker run -it --rm --privileged \
-     $USE_ALL_GPUS \
+    --name "$CONTAINER_NAME" \
+    $(if [ "$GPUS" -gt 0 ]; then echo "--gpus all"; fi) \
     -p $HTTP:8080 \
     $ADDITIONAL \
     -e WINSLOW_STORAGE_TYPE=$STORAGE_TYPE \
     -e WINSLOW_STORAGE_PATH=$STORAGE_PATH \
     -e WINSLOW_WORK_DIRECTORY=$WORKDIR \
     -e "WINSLOW_NODE_NAME=$NODE_NAME" \
+    $(if [ "$NODE_TYPE" == "observer" ]; then echo "-e WINSLOW_NO_STAGE_EXECUTION=1"; fi) \
     -v /var/run/docker.sock:/var/run/docker.sock \
     $IMAGE \
     $PARAMS
